requirements.txt
# Basic dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Level 1 - PDF reading
pypdf>=3.0.0

# AI/LLM - Gemini
google-generativeai>=0.8.0

# Level 2 - MCP server (FastMCP 2.0)
fastmcp>=2.0.0

# HTTP client
httpx>=0.25.0

# Data validation
pydantic>=2.0.0

# Type hints support
typing-extensions>=4.8.0

# Optional - for UI (Level 1)
streamlit>=1.28.0